requirements.txt
Babel==2.15.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
python-dotenv==1.0.1
speaklater==1.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
WTForms==3.1.2

marshmallow~=3.21.3
alembic~=1.13.1